Reorganize the benchmarks to build 1 executable per benchmark unit instead of a giant binary with everything inside
