+++ benches/bench-hashtable-op-set.cpp
@@ -22,7 +22,8 @@
 
 #include "../tests/support.h"
 
-#include "bench-support.h"
+#include "benchmark-program.hpp"
+#include "benchmark-support.hpp"
 
 // Set the generator to use
 #define KEYSET_GENERATOR_METHOD     TEST_SUPPORT_RANDOM_KEYS_GEN_FUNC_RANDOM_STR_MAX_LENGTH
@@ -66,7 +67,7 @@ public:
         this->_requested_keyset_size = (uint64_t) (((double) state.range(0)) * requested_load_factor);
 
         if (state.thread_index() == 0) {
-            if (bench_support_check_if_too_many_threads_per_core(
+            if (BenchmarkSupport::CheckIfTooManyThreadsPerCore(
                     state.threads(),
                     BENCHES_MAX_THREADS_PER_CORE)) {
                 sprintf(error_message, "Too many threads per core, max allowed <%d>", BENCHES_MAX_THREADS_PER_CORE);
@@ -117,7 +118,7 @@ public:
         }
 
         if (this->_hashtable != nullptr) {
-            bench_support_collect_hashtable_stats_and_update_state(
+            BenchmarkSupport::CollectHashtableStatsAndUpdateState(
                     (benchmark::State&)state, this->_hashtable);
 
             // Free the stoarge
@@ -207,7 +208,7 @@ public:
         this->_requested_keyset_size = (uint64_t) (((double) state.range(0)) * requested_load_factor);
 
         if (state.thread_index() == 0) {
-            if (bench_support_check_if_too_many_threads_per_core(
+            if (BenchmarkSupport::CheckIfTooManyThreadsPerCore(
                     state.threads(),
                     BENCHES_MAX_THREADS_PER_CORE)) {
                 sprintf(error_message, "Too many threads per core, max allowed <%d>", BENCHES_MAX_THREADS_PER_CORE);
@@ -311,7 +312,7 @@ public:
         }
 
         if (this->_hashtable != nullptr) {
-            bench_support_collect_hashtable_stats_and_update_state(
+            BenchmarkSupport::CollectHashtableStatsAndUpdateState(
                     (benchmark::State&)state, this->_hashtable);
 
             // Free the stoarge
@@ -376,7 +377,7 @@ BENCHMARK_DEFINE_F(HashtableOpSetUpdateFixture, hashtable_op_set_update)(benchma
 BENCHMARK_REGISTER_F(HashtableOpSetInsertFixture, hashtable_op_set_insert)
         ->ArgsProduct({
                               { 0x0000FFFFu, 0x000FFFFFu, 0x001FFFFFu, 0x007FFFFFu, 0x00FFFFFFu, 0x01FFFFFFu, 0x07FFFFFFu,
-                                0x0FFFFFFFu, 0x1FFFFFFFu, 0x3FFFFFFFu, 0x7FFFFFFFu, 0x7FFFFFFFu },
+                                0x0FFFFFFFu, 0x1FFFFFFFu, 0x3FFFFFFFu, 0x7FFFFFFFu },
                               { 50, 75 },
                       })
         ->ThreadRange(1, utils_cpu_count())
@@ -387,7 +388,7 @@ BENCHMARK_REGISTER_F(HashtableOpSetInsertFixture, hashtable_op_set_insert)
 BENCHMARK_REGISTER_F(HashtableOpSetUpdateFixture, hashtable_op_set_update)
         ->ArgsProduct({
                               { 0x0000FFFFu, 0x000FFFFFu, 0x001FFFFFu, 0x007FFFFFu, 0x00FFFFFFu, 0x01FFFFFFu, 0x07FFFFFFu,
-                                0x0FFFFFFFu, 0x1FFFFFFFu, 0x3FFFFFFFu, 0x7FFFFFFFu, 0x7FFFFFFFu },
+                                0x0FFFFFFFu, 0x1FFFFFFFu, 0x3FFFFFFFu, 0x7FFFFFFFu },
                               { 50, 75 },
                       })
         ->ThreadRange(1, utils_cpu_count())
